re #122 remove useless system tests
